CUMULUS-760 Re-organizing location of test utils
